test/requirements.txt
kafka-logging-handler==0.2.3
kafka-python==2.0.1
